% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_AverageBy.R
\name{AverageBy}
\alias{AverageBy}
\title{Average a numeric matrix or data.frame by columns}
\usage{
AverageBy(cts, by)
}
\arguments{
\item{cts}{a numeric matrix or data.frame}

\item{by}{a named list with column numbers to be averaged, colname will be the respective list name}
}
\description{
Function takes a named list with column numbers to be averaged. Unpaired columns are simply returned as-is.
}
\details{

}
\examples{
cts <- lapply(seq(1,5), function(x) rnorm(10)) \%>\% do.call(cbind, .)
colnames(cts) <- c("A_rep1", "B_rep1", "A_rep2", "C_rep1", "B_rep2")
rownames(cts) <- paste0("Gene", seq(1,nrow(cts)))
by <- sapply(unique(gsub("_rep.*", "", colnames(cts))), function(x) grep(paste0("^", x, "_rep"), colnames(cts)))
AverageBy(cts,by)

}
\author{
Alexander Toenges
}
