% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_gosty.R
\name{gosty}
\alias{gosty}
\title{Functional enrichment of a gene list with gprofiler2}
\usage{
gosty(
  InputGenes,
  BackgroundGenes = NULL,
  Species = "mmusculus",
  Sources = c("KEGG", "REAC"),
  FDR.threshold = 0.05,
  ...
)
}
\arguments{
\item{InputGenes}{vector with gene names (HGNC/MGI/Ensembl)}

\item{BackgroundGenes}{vector with gene names (HGNC/MGI/Ensembl)}

\item{Species}{mmusculus/hsapiens etc...}

\item{Sources}{the databases to use, default is c("KEGG", "REAC")}

\item{FDR.threshold}{FDR threshold for significance, if NULL return all hits}

\item{...}{further arguments for \code{gprofiler2::gost()}}
}
\description{
Function minimally expects a vector of genes which will be used for functional enrichment analysis
with GOST (g:profiler2). Optionally (and recommended) is a list of background genes.
One can specify organism and databases to scan against. See `?gost`, this here is just a simple wrapper to bring
output into a more parsable format.
}
\examples{

gosty(InputGenes = c("Ets1", "Cebpa", "Cd14", "Gapdh", "Spi1", "Cebpe"))

}
\author{
Alexander Toenges
}
