% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_makeAllContrasts.R
\name{makeAllContrasts}
\alias{makeAllContrasts}
\title{Make all possible unique pairwise contrasts}
\usage{
makeAllContrasts(group, delim = "_vs_", deseq2 = FALSE, name = "group")
}
\arguments{
\item{group}{vector with group names}

\item{delim}{a string to be used as delimiter for the names, Default is "_vs_", e.g. group1_vs_group2}

\item{deseq2}{logical, whether to return contrasts as a name list in DESeq2-ish style, like c("group", "A", "B")}

\item{name}{if deseq2, then the "group" name}
}
\description{
Expecting a vector with group names, e.g. `c("group1", "group2", "group3")` this function will
then create all unique pairwise contrasts. 
By default the output is a matrix/array in the limma-ish style of `makeContrasts()`.
If `deseq2=TRUE` it returns a named list with DESeq2-ish contrasts that can be used with the contrasts argument
of `results()` and `lfcShrink()`.
The group vector should not contain hyphen!
}
\details{

}
\examples{
makeAllContrasts(group = c("A", "B", "C"))
makeAllContrasts(group = c("A", "B", "C"), deseq2 = TRUE)

}
\author{
Alexander Toenges
}
