% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_MAplotFromCts.R
\name{MAplotFromCts}
\alias{MAplotFromCts}
\title{Pairwise MA-plots using smoothScatter for QC purposes}
\usage{
MAplotFromCts(
  x,
  y,
  Plot.title = "",
  cts.arelog2 = FALSE,
  type = "base",
  abline0 = FALSE,
  abline2 = NULL,
  use.loess = FALSE,
  ...
)
}
\arguments{
\item{x}{counts of sample1}

\item{y}{counts of sample2}

\item{cts.arelog2}{logical, whether counts are already on log2-scale}

\item{type}{the type of the plot, currently only smoothScatter from base is available, later a ggplot version will come}

\item{abline0}{logical, whether to add a horizontal ablone at y=0}

\item{abline2}{numeric, add horizontal ablines at +/- abline2 value}

\item{use.loess}{logical, whether to add a y ~ x  loess fitted line}

\item{...}{further arguments passed to smoothMAplot()}
}
\description{
Given two vectors of normalized counts make an MA-plot to see how normalization performed
and whether things look as expected.
}
\details{
the plot will be directly plotted to devise, so best would be to capture function directly to disk
}
\examples{
dds  <- estimateSizeFactors(makeExampleDESeqDataSet())
x <- counts(dds,normalized=TRUE)[,1]
y <- counts(dds,normalized=TRUE)[,2]
MAplotFromCts(x = x,y = y, use.loess=TRUE, Y.limits=c(-2,2))
}
\author{
Alexander Toenges
}
