% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_MostVariableRows.R
\name{MostVariableRows}
\alias{MostVariableRows}
\title{Get most variable rows from a numeric matrix#'}
\usage{
MostVariableRows(counts, ntop = 1000, type = c("rv", "mgv"))
}
\arguments{
\item{counts}{numeric matrix}

\item{ntop}{number of most most variable rows to return}

\item{type}{c("rv", "mgv") rv: selection by rowVars and mgv: selection based on 
the scran function \code{modelGeneVar} followed by \code{getTopHVGs} (take the top ntop) and return rownames}
}
\description{
Find most variable rows of a numeric matrix. If rownames are present returns those,
if not returns row index.
}
\details{

}
\examples{
counts <-lapply(seq(1,4), function(x) rnorm(20)) \%>\% do.call(cbind, .)
MostVariableRows(counts, 10)

}
\author{
Alexander Toenges
}
