% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_NormOffsetsFromTxLength.R
\name{NormOffsetsFromTxLength}
\alias{NormOffsetsFromTxLength}
\title{Combine AveTxLen with norm.factors into offsets}
\usage{
NormOffsetsFromTxLength(cts, len, norm_method = "TMM")
}
\arguments{
\item{cts}{count matrix}

\item{len}{length matrix}

\item{norm_method}{norm method from edgeR::calcNormFactors}
}
\description{
Wrapper around https://bioconductor.org/packages/release/bioc/vignettes/tximport/inst/doc/tximport.html#edger
Combine edgeR normalization factors with average transcript length and obtain norm. offsets for downstream use.
}
\details{

}
\author{
Alexander Toenges
}
