% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_res_edgeR.R
\name{res_edgeR}
\alias{res_edgeR}
\title{Automate the glmQLFit/glmTreat functions from edgeR}
\usage{
res_edgeR(fit, contrast = NULL, coef = NULL, lfc = log2(1.2), ID = "Gene")
}
\arguments{
\item{fit}{DGEGLM, output of glmQLFit}

\item{contrast}{a numeric contrast so a column from edgeR::makeContrasts()}

\item{coef}{a coefficient to test, ignored it contrast is set}

\item{lfc}{a lfc threshold for glmTreat, if 0 falls back to glmQLTest}

\item{ID}{the name of the column that will store the genes/regions/identifiers for each row}
}
\description{
Function assumes a DGEGLM object from glmQLFit and will then produce DE results based on a coef or contrast.
}
\examples{
nlibs <- 4
ngenes <- 1000
dispersion.true <- 1/rchisq(ngenes, df=10)
design <- model.matrix(~factor(c(1,1,2,2)))
y <- rnbinom(ngenes*nlibs,mu=20,size=1/dispersion.true)
y <- matrix(y,ngenes,nlibs)
rownames(y) <- paste0("Gene", seq(1,nrow(y)))
d <- DGEList(y)
d <- calcNormFactors(d)
d <- estimateDisp(d, design)
fit <- glmQLFit(d, design)
head(res_edgeR(fit=fit,coef=2))

}
\author{
Alexander Toenges
}
