% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_ScaleByQuantile.R
\name{ScaleByQuantile}
\alias{ScaleByQuantile}
\title{Winsorize a count matrix using fixed quantiles by row}
\usage{
ScaleByQuantile(Counts, lower = 0, upper = 1)
}
\arguments{
\item{Counts}{a matrix or dataframe with the count data}

\item{lower}{bottom quantile}

\item{upper}{top quantile}
}
\description{
Per row trim the values beyond the specified quantiles to that quantile,
can be useful e.g. when plotting heatmaps to limit influence of outliers
Takes a numeric matrix or data.frame and returns one with the winsorized values.
}
\examples{

Counts <- sapply(seq(1,10), function(x) rnorm(1000,100))
ScaleByQuantile(Counts=Counts,lower=0.1,upper=0.9)
 
}
\author{
Alexander Toenges
}
