% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_string2chr.R
\name{string2chr}
\alias{string2chr}
\title{Convert strsplit to data.frame}
\usage{
string2chr(
  x,
  pattern = ":|-",
  chr.name = "chr",
  start.name = "start",
  end.name = "end"
)
}
\arguments{
\item{x}{input string(s)}

\item{pattern}{pattern to split string at. Can be regexed.}

\item{chr.name}{name for chr}

\item{start.name}{name for start}

\item{end.name}{name for end}
}
\description{
Assuming coordinates such as chr:start-end, function splits at delimiters and produces
data.frame output. Uses `data.table::tstrsplit` internally.
}
\details{

}
\examples{
strs <- c("chr1:1-10", "chr2:2-20")
string2chr(strs, ":|-")

}
\author{
Alexander Toenges
}
