% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_Permut_GRanges.R,
%   R/src_Permut_GRanges_devel.R
\name{Permut_GRanges}
\alias{Permut_GRanges}
\title{A function to generate permutation-based p-values based on overlap between GRanges objects}
\usage{
Permut_GRanges(
  featureset,
  foreground,
  background,
  restrictset = NULL,
  nperm = 5000,
  conf.level = 0.95,
  return.values = FALSE
)

Permut_GRanges(
  featureset,
  foreground,
  background,
  restrictset = NULL,
  nperm = 5000,
  conf.level = 0.95,
  return.values = FALSE
)
}
\arguments{
\item{featureset}{GRanges object to overlap with the below objects, see details.}

\item{foreground}{GRanges object containing regions that will be used to generate 
the observed overlaps with the featureset set}

\item{background}{GRanges object containing regions that will be used to generate 
the permutated/random overlaps with the featureset set}

\item{restrictset}{GRanges to restrict overlapping operations between featureset and fore/background to.
This could e.g. be TAD boundaries and only overlaps between featureset and fore/background within the same TAD would be counted.
Features outside TADs are ignored.}

\item{nperm}{number of permutations. Smallest possible p-value is \code{1/(nperm+1)}}

\item{conf.level}{the confidence level to calculate}

\item{return.values}{logical, whether to return the number of overlaps between permuted backgrounds and featureset
as part of the output list. Can be useful for plotting/visualization, is not ordered though.}
}
\description{
Output is a named list with p-values, observed overlap and confidence intervals for 
the permutation-based overlap test.

Output is a named list with p-values, observed overlap and confidence intervals for 
the permutation-based overlap test.
}
\details{
Underlying question is whether the overlap for a given set of features (featureset), e.g.
ChIP-seq peaks is greater between a given foreground set, e.g. certain TSS, compared to a 
background, e.g. all annotated TSS. The permutation-based test performs overlap counting between
the foreground and permutated background (number of permutations by default 5000). Pvalues are then calculated
which test whether the background overlaps are equal or more extreme than the foreground overlaps.
The smallest possible pvalue is defined by 1/(nperm+1).
Optionally one can provide a restrictset, e.g. TAD coordinates, so only intersections between features
are counted that overlap the same restrictset interval.

Underlying question is whether the overlap for a given set of features (featureset), e.g.
ChIP-seq peaks is greater between a given foreground set, e.g. certain TSS, compared to a 
background, e.g. all annotated TSS. The permutation-based test performs overlap counting between
the foreground and permutated background (number of permutations by default 5000). Pvalues are then calculated
which test whether the background overlaps are equal or more extreme than the foreground overlaps.
The smallest possible pvalue is defined by 1/(nperm+1).
Optionally one can provide a restrictset, e.g. TAD coordinates, so only intersections between features
are counted that overlap the same restrictset interval.
}
\examples{
library(airway)
data(airway)

background=reduce(unlist(rowRanges(airway)[c(1:500)]))
set.seed(2021)
foreground=background[sample(1:length(background), 50, replace=FALSE)]
featureset=foreground
Permut_GRanges(featureset=featureset,foreground=foreground,background=background)

library(airway)
data(airway)

background=reduce(unlist(rowRanges(airway)[c(1:500)]))
set.seed(2021)
foreground=background[sample(1:length(background), 50, replace=FALSE)]
featureset=foreground
Permut_GRanges(featureset=featureset,foreground=foreground,background=background)

}
\author{
Alexander Toenges

Alexander Toenges
}
